change datebase structure
