+++ DDL&insert/ebook_dbo_t_booklist.sql
@@ -1,4 +1,4 @@
-INSERT INTO ebook.dbo.bookBasic (title, author, publisher, price, ID, stock, ISBN) VALUES ('尼采哲学经典(套装共5册)', '尼采,刘琦,余鸿荣', '哈尔滨出版社', 9.99, '1901075145', 10, '9787548420460');
-INSERT INTO ebook.dbo.bookBasic (title, author, publisher, price, ID, stock, ISBN) VALUES ('《三体》导读', '詹琰,路姜波', '天津人民出版社', 6.89, '1900685849', 6, '9787201108926');
-INSERT INTO ebook.dbo.bookBasic (title, author, publisher, price, ID, stock, ISBN) VALUES ('三体3：死神永生', '刘慈欣', '读客文化', 24.74, '1901104643', 150, '9787229030933');
-INSERT INTO ebook.dbo.bookBasic (title, author, publisher, price, ID, stock, ISBN) VALUES ('三体1：地球往事', '刘慈欣', '读客文化', 16.99, '1901104639', 14, '9787536692930');
+INSERT INTO ebook.dbo.books (title, author, publisher, price, ID, stock, ISBN) VALUES ('尼采哲学经典(套装共5册)', '尼采,刘琦,余鸿荣', '哈尔滨出版社', 9.99, '1901075145', 10, '9787548420460');
+INSERT INTO ebook.dbo.books (title, author, publisher, price, ID, stock, ISBN) VALUES ('《三体》导读', '詹琰,路姜波', '天津人民出版社', 6.89, '1900685849', 6, '9787201108926');
+INSERT INTO ebook.dbo.books (title, author, publisher, price, ID, stock, ISBN) VALUES ('三体3：死神永生', '刘慈欣', '读客文化', 24.74, '1901104643', 150, '9787229030933');
+INSERT INTO ebook.dbo.books (title, author, publisher, price, ID, stock, ISBN) VALUES ('三体1：地球往事', '刘慈欣', '读客文化', 16.99, '1901104639', 14, '9787536692930');
