+++ src/main/java/ebookBackend/controller/BookController.java
@@ -1,9 +1,8 @@
 package ebookBackend.controller;
 
-import ebookBackend.entity.Book;
-import ebookBackend.entity.BookBasic;
-import ebookBackend.service.BookBasicService;
-import ebookBackend.service.BookDetailService;
+//import ebookBackend.entity.Book;
+import ebookBackend.entity.Books;
+import ebookBackend.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,26 +10,40 @@ import java.util.List;
 
 
 @RestController
-@RequestMapping("/getBook")
+@RequestMapping("/book")
 public class BookController {
 
     @Autowired
-    BookBasicService bookBasicService;
+    BookService bookService;
 
     @RequestMapping(path = "/all", method = RequestMethod.GET)
     @ResponseBody
-    public List<BookBasic> getList() {
-        List<BookBasic> list = bookBasicService.listAll();
+    public List<Books> getList() {
+        List<Books> list = bookService.listAll();
         return list;
     }
 
-    @Autowired
-    BookDetailService bookDetailService;
-
     @RequestMapping(path = "/information", method = RequestMethod.GET)
     @ResponseBody
-    public Book information(@RequestParam String ID) {
-//        Book book = new Book(bookBasicService.get(ID),bookDetailService.get(ID));
-        return new Book(bookBasicService.get(ID),bookDetailService.get(ID));
+    public Books information(@RequestParam String ID) {
+        return bookService.get(ID);
+    }
+
+    @RequestMapping(path = "/update", method = RequestMethod.POST)
+    @ResponseBody
+    public int update(@RequestBody Books book) {
+        return bookService.update(book);
+    }
+
+    @RequestMapping(path = "/new", method = RequestMethod.POST)
+    @ResponseBody
+    public int newBook(@RequestBody Books book) {
+        return bookService.insert(book);
+    }
+
+    @RequestMapping(path = "/delete", method = RequestMethod.GET)
+    @ResponseBody
+    public int deleteBook(@RequestParam String id) {
+        return bookService.delete(id);
     }
 }
