+++ src/main/java/ebookBackend/controller/OrderController.java
@@ -3,14 +3,10 @@ package ebookBackend.controller;
 //import ebookBackend.entity.User;
 
 import ebookBackend.entity.Order;
-import ebookBackend.entity.User;
+import ebookBackend.entity.OrderWithItems;
 import ebookBackend.service.OrderService;
-import ebookBackend.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/order")
@@ -40,9 +36,9 @@ public class OrderController {
 //    }
 
     @RequestMapping(value = "/insert", method = RequestMethod.POST)
-    public Order insertOrder(@RequestBody Order order) {
-
-        return  orderService.makeOrder(userID);
+    public Order insertOrder(@RequestBody OrderWithItems order) {
+//        order.generateItemList();
+        return  orderService.makeOrder(order);
     }
