--- src/main/java/ebookBackend/dao/BookBasicMapper.java
@@ -1,64 +1,0 @@
-package ebookBackend.dao;
-
-import ebookBackend.entity.BookBasic;
-import ebookBackend.entity.BookBasicExample;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface BookBasicMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int countByExample(BookBasicExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int deleteByExample(BookBasicExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int insert(BookBasic record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int insertSelective(BookBasic record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    List<BookBasic> selectByExample(BookBasicExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int updateByExampleSelective(@Param("record") BookBasic record, @Param("example") BookBasicExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookBasic
-     *
-     * @mbggenerated
-     */
-    int updateByExample(@Param("record") BookBasic record, @Param("example") BookBasicExample example);
-}
