--- src/main/java/ebookBackend/dao/BookDetailMapper.java
@@ -1,64 +1,0 @@
-package ebookBackend.dao;
-
-import ebookBackend.entity.BookDetail;
-import ebookBackend.entity.BookDetailExample;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface BookDetailMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int countByExample(BookDetailExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int deleteByExample(BookDetailExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int insert(BookDetail record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int insertSelective(BookDetail record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    List<BookDetail> selectByExample(BookDetailExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int updateByExampleSelective(@Param("record") BookDetail record, @Param("example") BookDetailExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    int updateByExample(@Param("record") BookDetail record, @Param("example") BookDetailExample example);
-}
