+++ src/main/java/ebookBackend/dao/BooksMapper.java
@@ -1,0 +1,96 @@
+package ebookBackend.dao;
+
+import ebookBackend.entity.Books;
+import ebookBackend.entity.BooksExample;
+import java.util.List;
+import org.apache.ibatis.annotations.Param;
+
+public interface BooksMapper {
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int countByExample(BooksExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int deleteByExample(BooksExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int deleteByPrimaryKey(String id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int insert(Books record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int insertSelective(Books record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    List<Books> selectByExample(BooksExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    Books selectByPrimaryKey(String id);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int updateByExampleSelective(@Param("record") Books record, @Param("example") BooksExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int updateByExample(@Param("record") Books record, @Param("example") BooksExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int updateByPrimaryKeySelective(Books record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table books
+     *
+     * @mbggenerated
+     */
+    int updateByPrimaryKey(Books record);
+}
