+++ src/main/java/ebookBackend/dao/OrderMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Param;
 public interface OrderMapper {
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -16,7 +16,7 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -24,7 +24,15 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
+     *
+     * @mbggenerated
+     */
+    int deleteByPrimaryKey(String orderid);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -32,7 +40,7 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -40,7 +48,7 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -48,7 +56,15 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
+     *
+     * @mbggenerated
+     */
+    Order selectByPrimaryKey(String orderid);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -56,9 +72,25 @@ public interface OrderMapper {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
     int updateByExample(@Param("record") Order record, @Param("example") OrderExample example);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table orders
+     *
+     * @mbggenerated
+     */
+    int updateByPrimaryKeySelective(Order record);
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method corresponds to the database table orders
+     *
+     * @mbggenerated
+     */
+    int updateByPrimaryKey(Order record);
 }
