--- src/main/java/ebookBackend/entity/BookBasic.java
@@ -1,229 +1,0 @@
-package ebookBackend.entity;
-
-import java.math.BigDecimal;
-
-public class BookBasic {
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.title
-     *
-     * @mbggenerated
-     */
-    private String title;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.author
-     *
-     * @mbggenerated
-     */
-    private String author;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.publisher
-     *
-     * @mbggenerated
-     */
-    private String publisher;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.price
-     *
-     * @mbggenerated
-     */
-    private BigDecimal price;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.ID
-     *
-     * @mbggenerated
-     */
-    private String id;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.stock
-     *
-     * @mbggenerated
-     */
-    private Integer stock;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column bookBasic.ISBN
-     *
-     * @mbggenerated
-     */
-    private String isbn;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.title
-     *
-     * @return the value of bookBasic.title
-     *
-     * @mbggenerated
-     */
-    public String getTitle() {
-        return title;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.title
-     *
-     * @param title the value for bookBasic.title
-     *
-     * @mbggenerated
-     */
-    public void setTitle(String title) {
-        this.title = title == null ? null : title.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.author
-     *
-     * @return the value of bookBasic.author
-     *
-     * @mbggenerated
-     */
-    public String getAuthor() {
-        return author;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.author
-     *
-     * @param author the value for bookBasic.author
-     *
-     * @mbggenerated
-     */
-    public void setAuthor(String author) {
-        this.author = author == null ? null : author.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.publisher
-     *
-     * @return the value of bookBasic.publisher
-     *
-     * @mbggenerated
-     */
-    public String getPublisher() {
-        return publisher;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.publisher
-     *
-     * @param publisher the value for bookBasic.publisher
-     *
-     * @mbggenerated
-     */
-    public void setPublisher(String publisher) {
-        this.publisher = publisher == null ? null : publisher.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.price
-     *
-     * @return the value of bookBasic.price
-     *
-     * @mbggenerated
-     */
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.price
-     *
-     * @param price the value for bookBasic.price
-     *
-     * @mbggenerated
-     */
-    public void setPrice(BigDecimal price) {
-        this.price = price;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.ID
-     *
-     * @return the value of bookBasic.ID
-     *
-     * @mbggenerated
-     */
-    public String getId() {
-        return id;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.ID
-     *
-     * @param id the value for bookBasic.ID
-     *
-     * @mbggenerated
-     */
-    public void setId(String id) {
-        this.id = id == null ? null : id.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.stock
-     *
-     * @return the value of bookBasic.stock
-     *
-     * @mbggenerated
-     */
-    public Integer getStock() {
-        return stock;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.stock
-     *
-     * @param stock the value for bookBasic.stock
-     *
-     * @mbggenerated
-     */
-    public void setStock(Integer stock) {
-        this.stock = stock;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column bookBasic.ISBN
-     *
-     * @return the value of bookBasic.ISBN
-     *
-     * @mbggenerated
-     */
-    public String getIsbn() {
-        return isbn;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column bookBasic.ISBN
-     *
-     * @param isbn the value for bookBasic.ISBN
-     *
-     * @mbggenerated
-     */
-    public void setIsbn(String isbn) {
-        this.isbn = isbn == null ? null : isbn.trim();
-    }
-}
