--- src/main/java/ebookBackend/entity/BookDetailExample.java
@@ -1,702 +1,0 @@
-package ebookBackend.entity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class BookDetailExample {
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    protected String orderByClause;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    protected boolean distinct;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    protected List<Criteria> oredCriteria;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public BookDetailExample() {
-        oredCriteria = new ArrayList<Criteria>();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public void setOrderByClause(String orderByClause) {
-        this.orderByClause = orderByClause;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public String getOrderByClause() {
-        return orderByClause;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public void setDistinct(boolean distinct) {
-        this.distinct = distinct;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public boolean isDistinct() {
-        return distinct;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public List<Criteria> getOredCriteria() {
-        return oredCriteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public void or(Criteria criteria) {
-        oredCriteria.add(criteria);
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public Criteria or() {
-        Criteria criteria = createCriteriaInternal();
-        oredCriteria.add(criteria);
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public Criteria createCriteria() {
-        Criteria criteria = createCriteriaInternal();
-        if (oredCriteria.size() == 0) {
-            oredCriteria.add(criteria);
-        }
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    protected Criteria createCriteriaInternal() {
-        Criteria criteria = new Criteria();
-        return criteria;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public void clear() {
-        oredCriteria.clear();
-        orderByClause = null;
-        distinct = false;
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    protected abstract static class GeneratedCriteria {
-        protected List<Criterion> criteria;
-
-        protected GeneratedCriteria() {
-            super();
-            criteria = new ArrayList<Criterion>();
-        }
-
-        public boolean isValid() {
-            return criteria.size() > 0;
-        }
-
-        public List<Criterion> getAllCriteria() {
-            return criteria;
-        }
-
-        public List<Criterion> getCriteria() {
-            return criteria;
-        }
-
-        protected void addCriterion(String condition) {
-            if (condition == null) {
-                throw new RuntimeException("Value for condition cannot be null");
-            }
-            criteria.add(new Criterion(condition));
-        }
-
-        protected void addCriterion(String condition, Object value, String property) {
-            if (value == null) {
-                throw new RuntimeException("Value for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value));
-        }
-
-        protected void addCriterion(String condition, Object value1, Object value2, String property) {
-            if (value1 == null || value2 == null) {
-                throw new RuntimeException("Between values for " + property + " cannot be null");
-            }
-            criteria.add(new Criterion(condition, value1, value2));
-        }
-
-        public Criteria andIdIsNull() {
-            addCriterion("ID is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdIsNotNull() {
-            addCriterion("ID is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdEqualTo(String value) {
-            addCriterion("ID =", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdNotEqualTo(String value) {
-            addCriterion("ID <>", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdGreaterThan(String value) {
-            addCriterion("ID >", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdGreaterThanOrEqualTo(String value) {
-            addCriterion("ID >=", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdLessThan(String value) {
-            addCriterion("ID <", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdLessThanOrEqualTo(String value) {
-            addCriterion("ID <=", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdLike(String value) {
-            addCriterion("ID like", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdNotLike(String value) {
-            addCriterion("ID not like", value, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdIn(List<String> values) {
-            addCriterion("ID in", values, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdNotIn(List<String> values) {
-            addCriterion("ID not in", values, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdBetween(String value1, String value2) {
-            addCriterion("ID between", value1, value2, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andIdNotBetween(String value1, String value2) {
-            addCriterion("ID not between", value1, value2, "id");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateIsNull() {
-            addCriterion("rate is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateIsNotNull() {
-            addCriterion("rate is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateEqualTo(Double value) {
-            addCriterion("rate =", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateNotEqualTo(Double value) {
-            addCriterion("rate <>", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateGreaterThan(Double value) {
-            addCriterion("rate >", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateGreaterThanOrEqualTo(Double value) {
-            addCriterion("rate >=", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateLessThan(Double value) {
-            addCriterion("rate <", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateLessThanOrEqualTo(Double value) {
-            addCriterion("rate <=", value, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateIn(List<Double> values) {
-            addCriterion("rate in", values, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateNotIn(List<Double> values) {
-            addCriterion("rate not in", values, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateBetween(Double value1, Double value2) {
-            addCriterion("rate between", value1, value2, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andRateNotBetween(Double value1, Double value2) {
-            addCriterion("rate not between", value1, value2, "rate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateIsNull() {
-            addCriterion("publishDate is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateIsNotNull() {
-            addCriterion("publishDate is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateEqualTo(String value) {
-            addCriterion("publishDate =", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateNotEqualTo(String value) {
-            addCriterion("publishDate <>", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateGreaterThan(String value) {
-            addCriterion("publishDate >", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateGreaterThanOrEqualTo(String value) {
-            addCriterion("publishDate >=", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateLessThan(String value) {
-            addCriterion("publishDate <", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateLessThanOrEqualTo(String value) {
-            addCriterion("publishDate <=", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateLike(String value) {
-            addCriterion("publishDate like", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateNotLike(String value) {
-            addCriterion("publishDate not like", value, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateIn(List<String> values) {
-            addCriterion("publishDate in", values, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateNotIn(List<String> values) {
-            addCriterion("publishDate not in", values, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateBetween(String value1, String value2) {
-            addCriterion("publishDate between", value1, value2, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andPublishdateNotBetween(String value1, String value2) {
-            addCriterion("publishDate not between", value1, value2, "publishdate");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsIsNull() {
-            addCriterion("words is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsIsNotNull() {
-            addCriterion("words is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsEqualTo(Double value) {
-            addCriterion("words =", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsNotEqualTo(Double value) {
-            addCriterion("words <>", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsGreaterThan(Double value) {
-            addCriterion("words >", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsGreaterThanOrEqualTo(Double value) {
-            addCriterion("words >=", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsLessThan(Double value) {
-            addCriterion("words <", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsLessThanOrEqualTo(Double value) {
-            addCriterion("words <=", value, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsIn(List<Double> values) {
-            addCriterion("words in", values, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsNotIn(List<Double> values) {
-            addCriterion("words not in", values, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsBetween(Double value1, Double value2) {
-            addCriterion("words between", value1, value2, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andWordsNotBetween(Double value1, Double value2) {
-            addCriterion("words not between", value1, value2, "words");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionIsNull() {
-            addCriterion("abstraction is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionIsNotNull() {
-            addCriterion("abstraction is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionEqualTo(String value) {
-            addCriterion("abstraction =", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionNotEqualTo(String value) {
-            addCriterion("abstraction <>", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionGreaterThan(String value) {
-            addCriterion("abstraction >", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionGreaterThanOrEqualTo(String value) {
-            addCriterion("abstraction >=", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionLessThan(String value) {
-            addCriterion("abstraction <", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionLessThanOrEqualTo(String value) {
-            addCriterion("abstraction <=", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionLike(String value) {
-            addCriterion("abstraction like", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionNotLike(String value) {
-            addCriterion("abstraction not like", value, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionIn(List<String> values) {
-            addCriterion("abstraction in", values, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionNotIn(List<String> values) {
-            addCriterion("abstraction not in", values, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionBetween(String value1, String value2) {
-            addCriterion("abstraction between", value1, value2, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andAbstractionNotBetween(String value1, String value2) {
-            addCriterion("abstraction not between", value1, value2, "abstraction");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueIsNull() {
-            addCriterion("catalogue is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueIsNotNull() {
-            addCriterion("catalogue is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueEqualTo(String value) {
-            addCriterion("catalogue =", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueNotEqualTo(String value) {
-            addCriterion("catalogue <>", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueGreaterThan(String value) {
-            addCriterion("catalogue >", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueGreaterThanOrEqualTo(String value) {
-            addCriterion("catalogue >=", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueLessThan(String value) {
-            addCriterion("catalogue <", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueLessThanOrEqualTo(String value) {
-            addCriterion("catalogue <=", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueLike(String value) {
-            addCriterion("catalogue like", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueNotLike(String value) {
-            addCriterion("catalogue not like", value, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueIn(List<String> values) {
-            addCriterion("catalogue in", values, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueNotIn(List<String> values) {
-            addCriterion("catalogue not in", values, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueBetween(String value1, String value2) {
-            addCriterion("catalogue between", value1, value2, "catalogue");
-            return (Criteria) this;
-        }
-
-        public Criteria andCatalogueNotBetween(String value1, String value2) {
-            addCriterion("catalogue not between", value1, value2, "catalogue");
-            return (Criteria) this;
-        }
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table bookDetail
-     *
-     * @mbggenerated do_not_delete_during_merge
-     */
-    public static class Criteria extends GeneratedCriteria {
-
-        protected Criteria() {
-            super();
-        }
-    }
-
-    /**
-     * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table bookDetail
-     *
-     * @mbggenerated
-     */
-    public static class Criterion {
-        private String condition;
-
-        private Object value;
-
-        private Object secondValue;
-
-        private boolean noValue;
-
-        private boolean singleValue;
-
-        private boolean betweenValue;
-
-        private boolean listValue;
-
-        private String typeHandler;
-
-        public String getCondition() {
-            return condition;
-        }
-
-        public Object getValue() {
-            return value;
-        }
-
-        public Object getSecondValue() {
-            return secondValue;
-        }
-
-        public boolean isNoValue() {
-            return noValue;
-        }
-
-        public boolean isSingleValue() {
-            return singleValue;
-        }
-
-        public boolean isBetweenValue() {
-            return betweenValue;
-        }
-
-        public boolean isListValue() {
-            return listValue;
-        }
-
-        public String getTypeHandler() {
-            return typeHandler;
-        }
-
-        protected Criterion(String condition) {
-            super();
-            this.condition = condition;
-            this.typeHandler = null;
-            this.noValue = true;
-        }
-
-        protected Criterion(String condition, Object value, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.typeHandler = typeHandler;
-            if (value instanceof List<?>) {
-                this.listValue = true;
-            } else {
-                this.singleValue = true;
-            }
-        }
-
-        protected Criterion(String condition, Object value) {
-            this(condition, value, null);
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue, String typeHandler) {
-            super();
-            this.condition = condition;
-            this.value = value;
-            this.secondValue = secondValue;
-            this.typeHandler = typeHandler;
-            this.betweenValue = true;
-        }
-
-        protected Criterion(String condition, Object value, Object secondValue) {
-            this(condition, value, secondValue, null);
-        }
-    }
-}
