+++ src/main/java/ebookBackend/entity/Books.java
@@ -1,0 +1,389 @@
+package ebookBackend.entity;
+
+import java.math.BigDecimal;
+
+public class Books {
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.ID
+     *
+     * @mbggenerated
+     */
+    private String id;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.title
+     *
+     * @mbggenerated
+     */
+    private String title;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.author
+     *
+     * @mbggenerated
+     */
+    private String author;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.publisher
+     *
+     * @mbggenerated
+     */
+    private String publisher;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.price
+     *
+     * @mbggenerated
+     */
+    private BigDecimal price;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.stock
+     *
+     * @mbggenerated
+     */
+    private Integer stock;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.ISBN
+     *
+     * @mbggenerated
+     */
+    private String isbn;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.rate
+     *
+     * @mbggenerated
+     */
+    private Double rate;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.words
+     *
+     * @mbggenerated
+     */
+    private Double words;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.publishDate
+     *
+     * @mbggenerated
+     */
+    private String publishdate;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.abstraction
+     *
+     * @mbggenerated
+     */
+    private String abstraction;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column books.catalogue
+     *
+     * @mbggenerated
+     */
+    private String catalogue;
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.ID
+     *
+     * @return the value of books.ID
+     *
+     * @mbggenerated
+     */
+    public String getId() {
+        return id;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.ID
+     *
+     * @param id the value for books.ID
+     *
+     * @mbggenerated
+     */
+    public void setId(String id) {
+        this.id = id == null ? null : id.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.title
+     *
+     * @return the value of books.title
+     *
+     * @mbggenerated
+     */
+    public String getTitle() {
+        return title;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.title
+     *
+     * @param title the value for books.title
+     *
+     * @mbggenerated
+     */
+    public void setTitle(String title) {
+        this.title = title == null ? null : title.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.author
+     *
+     * @return the value of books.author
+     *
+     * @mbggenerated
+     */
+    public String getAuthor() {
+        return author;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.author
+     *
+     * @param author the value for books.author
+     *
+     * @mbggenerated
+     */
+    public void setAuthor(String author) {
+        this.author = author == null ? null : author.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.publisher
+     *
+     * @return the value of books.publisher
+     *
+     * @mbggenerated
+     */
+    public String getPublisher() {
+        return publisher;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.publisher
+     *
+     * @param publisher the value for books.publisher
+     *
+     * @mbggenerated
+     */
+    public void setPublisher(String publisher) {
+        this.publisher = publisher == null ? null : publisher.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.price
+     *
+     * @return the value of books.price
+     *
+     * @mbggenerated
+     */
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.price
+     *
+     * @param price the value for books.price
+     *
+     * @mbggenerated
+     */
+    public void setPrice(BigDecimal price) {
+        this.price = price;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.stock
+     *
+     * @return the value of books.stock
+     *
+     * @mbggenerated
+     */
+    public Integer getStock() {
+        return stock;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.stock
+     *
+     * @param stock the value for books.stock
+     *
+     * @mbggenerated
+     */
+    public void setStock(Integer stock) {
+        this.stock = stock;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.ISBN
+     *
+     * @return the value of books.ISBN
+     *
+     * @mbggenerated
+     */
+    public String getIsbn() {
+        return isbn;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.ISBN
+     *
+     * @param isbn the value for books.ISBN
+     *
+     * @mbggenerated
+     */
+    public void setIsbn(String isbn) {
+        this.isbn = isbn == null ? null : isbn.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.rate
+     *
+     * @return the value of books.rate
+     *
+     * @mbggenerated
+     */
+    public Double getRate() {
+        return rate;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.rate
+     *
+     * @param rate the value for books.rate
+     *
+     * @mbggenerated
+     */
+    public void setRate(Double rate) {
+        this.rate = rate;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.words
+     *
+     * @return the value of books.words
+     *
+     * @mbggenerated
+     */
+    public Double getWords() {
+        return words;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.words
+     *
+     * @param words the value for books.words
+     *
+     * @mbggenerated
+     */
+    public void setWords(Double words) {
+        this.words = words;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.publishDate
+     *
+     * @return the value of books.publishDate
+     *
+     * @mbggenerated
+     */
+    public String getPublishdate() {
+        return publishdate;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.publishDate
+     *
+     * @param publishdate the value for books.publishDate
+     *
+     * @mbggenerated
+     */
+    public void setPublishdate(String publishdate) {
+        this.publishdate = publishdate == null ? null : publishdate.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.abstraction
+     *
+     * @return the value of books.abstraction
+     *
+     * @mbggenerated
+     */
+    public String getAbstraction() {
+        return abstraction;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.abstraction
+     *
+     * @param abstraction the value for books.abstraction
+     *
+     * @mbggenerated
+     */
+    public void setAbstraction(String abstraction) {
+        this.abstraction = abstraction == null ? null : abstraction.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column books.catalogue
+     *
+     * @return the value of books.catalogue
+     *
+     * @mbggenerated
+     */
+    public String getCatalogue() {
+        return catalogue;
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method sets the value of the database column books.catalogue
+     *
+     * @param catalogue the value for books.catalogue
+     *
+     * @mbggenerated
+     */
+    public void setCatalogue(String catalogue) {
+        this.catalogue = catalogue == null ? null : catalogue.trim();
+    }
+}
