+++ src/main/java/ebookBackend/entity/Items.java
@@ -9,7 +9,7 @@ public class Items {
      *
      * @mbggenerated
      */
-    private Integer orderid;
+    private String orderid;
 
     /**
      * This field was generated by MyBatis Generator.
@@ -43,7 +43,7 @@ public class Items {
      *
      * @mbggenerated
      */
-    public Integer getOrderid() {
+    public String getOrderid() {
         return orderid;
     }
 
@@ -55,8 +55,8 @@ public class Items {
      *
      * @mbggenerated
      */
-    public void setOrderid(Integer orderid) {
-        this.orderid = orderid;
+    public void setOrderid(String orderid) {
+        this.orderid = orderid == null ? null : orderid.trim();
     }
 
     /**
