+++ src/main/java/ebookBackend/entity/Order.java
@@ -5,7 +5,15 @@ import java.math.BigDecimal;
 public class Order {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column order.total
+     * This field corresponds to the database column orders.orderID
+     *
+     * @mbggenerated
+     */
+    private String orderid;
+
+    /**
+     * This field was generated by MyBatis Generator.
+     * This field corresponds to the database column orders.total
      *
      * @mbggenerated
      */
@@ -13,25 +21,41 @@ public class Order {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column order.userID
+     * This field corresponds to the database column orders.userID
      *
      * @mbggenerated
      */
     private String userid;
 
     /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column order.orderID
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column orders.orderID
+     *
+     * @return the value of orders.orderID
      *
      * @mbggenerated
      */
-    private Integer orderid;
+    public String getOrderid() {
+        return orderid;
+    }
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column order.total
+     * This method sets the value of the database column orders.orderID
      *
-     * @return the value of order.total
+     * @param orderid the value for orders.orderID
+     *
+     * @mbggenerated
+     */
+    public void setOrderid(String orderid) {
+        this.orderid = orderid == null ? null : orderid.trim();
+    }
+
+    /**
+     * This method was generated by MyBatis Generator.
+     * This method returns the value of the database column orders.total
+     *
+     * @return the value of orders.total
      *
      * @mbggenerated
      */
@@ -41,9 +65,9 @@ public class Order {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column order.total
+     * This method sets the value of the database column orders.total
      *
-     * @param total the value for order.total
+     * @param total the value for orders.total
      *
      * @mbggenerated
      */
@@ -53,9 +77,9 @@ public class Order {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column order.userID
+     * This method returns the value of the database column orders.userID
      *
-     * @return the value of order.userID
+     * @return the value of orders.userID
      *
      * @mbggenerated
      */
@@ -65,37 +89,13 @@ public class Order {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column order.userID
+     * This method sets the value of the database column orders.userID
      *
-     * @param userid the value for order.userID
+     * @param userid the value for orders.userID
      *
      * @mbggenerated
      */
     public void setUserid(String userid) {
         this.userid = userid == null ? null : userid.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column order.orderID
-     *
-     * @return the value of order.orderID
-     *
-     * @mbggenerated
-     */
-    public Integer getOrderid() {
-        return orderid;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column order.orderID
-     *
-     * @param orderid the value for order.orderID
-     *
-     * @mbggenerated
-     */
-    public void setOrderid(Integer orderid) {
-        this.orderid = orderid;
     }
 }
