+++ src/main/java/ebookBackend/entity/OrderExample.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class OrderExample {
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table order
+     * This field corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -15,7 +15,7 @@ public class OrderExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table order
+     * This field corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -23,7 +23,7 @@ public class OrderExample {
 
     /**
      * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table order
+     * This field corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -31,7 +31,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -41,7 +41,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -51,7 +51,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -61,7 +61,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -71,7 +71,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -81,7 +81,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -91,7 +91,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -101,7 +101,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -113,7 +113,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -127,7 +127,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -138,7 +138,7 @@ public class OrderExample {
 
     /**
      * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table order
+     * This method corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -150,7 +150,7 @@ public class OrderExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table order
+     * This class corresponds to the database table orders
      *
      * @mbggenerated
      */
@@ -193,6 +193,76 @@ public class OrderExample {
                 throw new RuntimeException("Between values for " + property + " cannot be null");
             }
             criteria.add(new Criterion(condition, value1, value2));
+        }
+
+        public Criteria andOrderidIsNull() {
+            addCriterion("orderID is null");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidIsNotNull() {
+            addCriterion("orderID is not null");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidEqualTo(String value) {
+            addCriterion("orderID =", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidNotEqualTo(String value) {
+            addCriterion("orderID <>", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidGreaterThan(String value) {
+            addCriterion("orderID >", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidGreaterThanOrEqualTo(String value) {
+            addCriterion("orderID >=", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidLessThan(String value) {
+            addCriterion("orderID <", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidLessThanOrEqualTo(String value) {
+            addCriterion("orderID <=", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidLike(String value) {
+            addCriterion("orderID like", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidNotLike(String value) {
+            addCriterion("orderID not like", value, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidIn(List<String> values) {
+            addCriterion("orderID in", values, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidNotIn(List<String> values) {
+            addCriterion("orderID not in", values, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidBetween(String value1, String value2) {
+            addCriterion("orderID between", value1, value2, "orderid");
+            return (Criteria) this;
+        }
+
+        public Criteria andOrderidNotBetween(String value1, String value2) {
+            addCriterion("orderID not between", value1, value2, "orderid");
+            return (Criteria) this;
         }
 
         public Criteria andTotalIsNull() {
@@ -324,71 +394,11 @@ public class OrderExample {
             addCriterion("userID not between", value1, value2, "userid");
             return (Criteria) this;
         }
-
-        public Criteria andOrderidIsNull() {
-            addCriterion("orderID is null");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidIsNotNull() {
-            addCriterion("orderID is not null");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidEqualTo(Integer value) {
-            addCriterion("orderID =", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidNotEqualTo(Integer value) {
-            addCriterion("orderID <>", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidGreaterThan(Integer value) {
-            addCriterion("orderID >", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidGreaterThanOrEqualTo(Integer value) {
-            addCriterion("orderID >=", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidLessThan(Integer value) {
-            addCriterion("orderID <", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidLessThanOrEqualTo(Integer value) {
-            addCriterion("orderID <=", value, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidIn(List<Integer> values) {
-            addCriterion("orderID in", values, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidNotIn(List<Integer> values) {
-            addCriterion("orderID not in", values, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidBetween(Integer value1, Integer value2) {
-            addCriterion("orderID between", value1, value2, "orderid");
-            return (Criteria) this;
-        }
-
-        public Criteria andOrderidNotBetween(Integer value1, Integer value2) {
-            addCriterion("orderID not between", value1, value2, "orderid");
-            return (Criteria) this;
-        }
     }
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table order
+     * This class corresponds to the database table orders
      *
      * @mbggenerated do_not_delete_during_merge
      */
@@ -401,7 +411,7 @@ public class OrderExample {
 
     /**
      * This class was generated by MyBatis Generator.
-     * This class corresponds to the database table order
+     * This class corresponds to the database table orders
      *
      * @mbggenerated
      */
