+++ src/main/java/ebookBackend/mapping/ItemsMapper.xml
@@ -6,7 +6,7 @@
       WARNING - @mbggenerated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    <result column="orderID" property="orderid" jdbcType="INTEGER" />
+    <result column="orderID" property="orderid" jdbcType="VARCHAR" />
     <result column="bookID" property="bookid" jdbcType="VARCHAR" />
     <result column="amount" property="amount" jdbcType="INTEGER" />
     <result column="value" property="value" jdbcType="DECIMAL" />
@@ -120,7 +120,7 @@
     -->
     insert into items (orderID, bookID, amount, 
       value)
-    values (#{orderid,jdbcType=INTEGER}, #{bookid,jdbcType=VARCHAR}, #{amount,jdbcType=INTEGER}, 
+    values (#{orderid,jdbcType=VARCHAR}, #{bookid,jdbcType=VARCHAR}, #{amount,jdbcType=INTEGER}, 
       #{value,jdbcType=DECIMAL})
   </insert>
   <insert id="insertSelective" parameterType="ebookBackend.entity.Items" >
@@ -145,7 +145,7 @@
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides="," >
       <if test="orderid != null" >
-        #{orderid,jdbcType=INTEGER},
+        #{orderid,jdbcType=VARCHAR},
       </if>
       <if test="bookid != null" >
         #{bookid,jdbcType=VARCHAR},
@@ -176,7 +176,7 @@
     update items
     <set >
       <if test="record.orderid != null" >
-        orderID = #{record.orderid,jdbcType=INTEGER},
+        orderID = #{record.orderid,jdbcType=VARCHAR},
       </if>
       <if test="record.bookid != null" >
         bookID = #{record.bookid,jdbcType=VARCHAR},
@@ -198,7 +198,7 @@
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
     update items
-    set orderID = #{record.orderid,jdbcType=INTEGER},
+    set orderID = #{record.orderid,jdbcType=VARCHAR},
       bookID = #{record.bookid,jdbcType=VARCHAR},
       amount = #{record.amount,jdbcType=INTEGER},
       value = #{record.value,jdbcType=DECIMAL}
