+++ src/main/java/ebookBackend/mapping/OrderMapper.xml
@@ -6,9 +6,9 @@
       WARNING - @mbggenerated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
+    <id column="orderID" property="orderid" jdbcType="VARCHAR" />
     <result column="total" property="total" jdbcType="DECIMAL" />
     <result column="userID" property="userid" jdbcType="VARCHAR" />
-    <result column="orderID" property="orderid" jdbcType="INTEGER" />
   </resultMap>
   <sql id="Example_Where_Clause" >
     <!--
@@ -81,7 +81,7 @@
       WARNING - @mbggenerated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    total, userID, orderID
+    orderID, total, userID
   </sql>
   <select id="selectByExample" resultMap="BaseResultMap" parameterType="ebookBackend.entity.OrderExample" >
     <!--
@@ -102,6 +102,24 @@
       order by ${orderByClause}
     </if>
   </select>
+  <select id="selectByPrimaryKey" resultMap="BaseResultMap" parameterType="java.lang.String" >
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    select 
+    <include refid="Base_Column_List" />
+    from orders
+    where orderID = #{orderid,jdbcType=VARCHAR}
+  </select>
+  <delete id="deleteByPrimaryKey" parameterType="java.lang.String" >
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    delete from orders
+    where orderID = #{orderid,jdbcType=VARCHAR}
+  </delete>
   <delete id="deleteByExample" parameterType="ebookBackend.entity.OrderExample" >
     <!--
       WARNING - @mbggenerated
@@ -112,13 +130,15 @@
       <include refid="Example_Where_Clause" />
     </if>
   </delete>
-  <insert id="insert" useGeneratedKeys="true" keyColumn="orderID" keyProperty="orderID" parameterType="ebookBackend.entity.Order" >
+  <insert id="insert" parameterType="ebookBackend.entity.Order" >
     <!--
       WARNING - @mbggenerated
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
-    insert into orders (total, userID)
-    values (#{total,jdbcType=DECIMAL}, #{userid,jdbcType=VARCHAR})
+    insert into orders (orderID, total, userID
+      )
+    values (#{orderid,jdbcType=VARCHAR}, #{total,jdbcType=DECIMAL}, #{userid,jdbcType=VARCHAR}
+      )
   </insert>
   <insert id="insertSelective" parameterType="ebookBackend.entity.Order" >
     <!--
@@ -127,25 +147,25 @@
     -->
     insert into orders
     <trim prefix="(" suffix=")" suffixOverrides="," >
+      <if test="orderid != null" >
+        orderID,
+      </if>
       <if test="total != null" >
         total,
       </if>
       <if test="userid != null" >
         userID,
       </if>
-      <if test="orderid != null" >
-        orderID,
-      </if>
     </trim>
     <trim prefix="values (" suffix=")" suffixOverrides="," >
+      <if test="orderid != null" >
+        #{orderid,jdbcType=VARCHAR},
+      </if>
       <if test="total != null" >
         #{total,jdbcType=DECIMAL},
       </if>
       <if test="userid != null" >
         #{userid,jdbcType=VARCHAR},
-      </if>
-      <if test="orderid != null" >
-        #{orderid,jdbcType=INTEGER},
       </if>
     </trim>
   </insert>
@@ -166,14 +186,14 @@
     -->
     update orders
     <set >
+      <if test="record.orderid != null" >
+        orderID = #{record.orderid,jdbcType=VARCHAR},
+      </if>
       <if test="record.total != null" >
         total = #{record.total,jdbcType=DECIMAL},
       </if>
       <if test="record.userid != null" >
         userID = #{record.userid,jdbcType=VARCHAR},
-      </if>
-      <if test="record.orderid != null" >
-        orderID = #{record.orderid,jdbcType=INTEGER},
       </if>
     </set>
     <if test="_parameter != null" >
@@ -186,11 +206,37 @@
       This element is automatically generated by MyBatis Generator, do not modify.
     -->
     update orders
-    set total = #{record.total,jdbcType=DECIMAL},
-      userID = #{record.userid,jdbcType=VARCHAR},
-      orderID = #{record.orderid,jdbcType=INTEGER}
+    set orderID = #{record.orderid,jdbcType=VARCHAR},
+      total = #{record.total,jdbcType=DECIMAL},
+      userID = #{record.userid,jdbcType=VARCHAR}
     <if test="_parameter != null" >
       <include refid="Update_By_Example_Where_Clause" />
     </if>
+  </update>
+  <update id="updateByPrimaryKeySelective" parameterType="ebookBackend.entity.Order" >
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    update orders
+    <set >
+      <if test="total != null" >
+        total = #{total,jdbcType=DECIMAL},
+      </if>
+      <if test="userid != null" >
+        userID = #{userid,jdbcType=VARCHAR},
+      </if>
+    </set>
+    where orderID = #{orderid,jdbcType=VARCHAR}
+  </update>
+  <update id="updateByPrimaryKey" parameterType="ebookBackend.entity.Order" >
+    <!--
+      WARNING - @mbggenerated
+      This element is automatically generated by MyBatis Generator, do not modify.
+    -->
+    update orders
+    set total = #{total,jdbcType=DECIMAL},
+      userID = #{userid,jdbcType=VARCHAR}
+    where orderID = #{orderid,jdbcType=VARCHAR}
   </update>
 </mapper>
